requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
fpdf>=2.7.0
Pillow>=10.0.0 